requirements.txt
arrow==0.15.4
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.19.1
pycparser==2.19
pylint==2.4.4
pymongo==3.9.0
python-dateutil==2.8.1
python-dotenv==0.10.5
PyYAML==5.3
six==1.13.0
slug==2.0
typed-ast==1.4.0
wrapt==1.11.2
xmltodict==0.12.0
../campus-shared-models
../campus-libs
